make mod steamodded independent, make functions not global to avoid conflicts
