+++ lovely.toml
@@ -1,0 +1,25 @@
+[manifest]
+version = "1.0.0"
+dump_lua = true
+priority = 11
+
+[[patches]]
+[patches.module]
+source = "FlushHotkeys.lua"
+before = "main.lua"
+name = "flushhotkeys"
+
+[[patches]]
+[patches.module]
+source = "FlushHotkeys.lua"
+before = "main.lua"
+name = "flushhotkeys"
+
+# Add core functionality
+[[patches]]
+[patches.copy]
+target = "main.lua"
+position = "append"
+sources = [
+    "init.lua",
+]
